+++ demos/Simple 02/Demo02_MoveCamera.py
@@ -56,9 +56,6 @@ class MoveCamera():
             self.lge.Quit()
 
         # mostramos info
-        fps = self.lge.GetFPS()
-        fps = "FPS: %07.2f" % fps
-
         mx, my = self.lge.GetMousePosition()
         mb1, mb2, mb3 = self.lge.GetMouseButtons()