+++ demos/Simple 02/Demo05_Colliders.py
@@ -66,9 +66,6 @@ class Colliders():
             self.lge.Quit()
 
         # mostramos info
-        fps = self.lge.GetFPS()
-        fps = "FPS: %07.2f" % fps
-
         mx, my = self.lge.GetMousePosition()
         mb1, mb2, mb3 = self.lge.GetMouseButtons()
 
@@ -108,7 +105,7 @@ class MiHeroe(Sprite):
         super().__init__(["heroe_idle_right", "heroe_idle_left", "heroe_run_right", "heroe_run_left"], (550, 346), "Heroe")
 
         # acceso al motor de juegos
-        self.lge = LittleGameEngine.GetLGE()
+        self.lge = self.GetLGE()
 
         # sus atributos
         self.SetOnEvents(LittleGameEngine.E_ON_UPDATE)
@@ -133,20 +130,20 @@ class MiHeroe(Sprite):
         if (self.lge.KeyPressed(LittleGameEngine.CONSTANTS.K_RIGHT)):
             x = x + pixels
             if (self.state != 2):
-                self.SetShape("heroe_run_right", 0)
+                self.SetShape("heroe_run_right")
                 self.state = 2
         elif(self.lge.KeyPressed(LittleGameEngine.CONSTANTS.K_LEFT)):
             x = x - pixels
             if (self.state != -2):
-                self.SetShape("heroe_run_left", 0)
+                self.SetShape("heroe_run_left")
                 self.state = -2
         elif(self.state == 2):
             if (self.state != 1):
-                self.SetShape("heroe_idle_right", 0)
+                self.SetShape("heroe_idle_right")
                 self.state = 1
         elif(self.state == -2):
             if (self.state != -1):
-                self.SetShape("heroe_idle_left", 0)
+                self.SetShape("heroe_idle_left")
                 self.state = -1
 
         if (self.lge.KeyPressed(LittleGameEngine.CONSTANTS.K_UP)):